Good morning!

+++ SwinIR/main_train_unet_plug-socket.py
@@ -42,7 +42,7 @@ def main():
     device = torch.device('cuda' if torch.cuda.is_available() else 'cpu')
 
     model = UNet(n_channels=3, n_classes=1, bilinear=True)
-    model.train()
+    model.train().double()
     model = model.to(device)
     criterion = nn.SmoothL1Loss()
     optimizer = torch.optim.AdamW(model.parameters(), lr=1e-4)
@@ -98,8 +98,8 @@ def main():
                     if output_channel == 1:
                         batch_y[idx_batch, 0, :, :] = cube_y_data[:, :, z_center]
 
-                batch_x = torch.from_numpy(batch_x).to(device)
-                batch_y = torch.from_numpy(batch_y).to(device)
+                batch_x = torch.from_numpy(batch_x).double().to(device)
+                batch_y = torch.from_numpy(batch_y).double().to(device)
 
                 optimizer.zero_grad()
                 y_hat = model(batch_x)
@@ -177,8 +177,8 @@ def main():
                     if output_channel == 1:
                         batch_y[idx_batch, 0, :, :] = cube_y_data[:, :, z_center]
 
-                batch_x = torch.from_numpy(batch_x).to(device)
-                batch_y = torch.from_numpy(batch_y).to(device)
+                batch_x = torch.from_numpy(batch_x).double().to(device)
+                batch_y = torch.from_numpy(batch_y).double().to(device)
                 
                 y_hat = model(batch_x)
                 loss = criterion(y_hat, batch_y)
